Move StorageClassDelegate back out of datastore subpackage.
This required a small change to class variables in StorageClass; it
turns out the previous approach was (weakly) vulnerable to the order
in which modules appear in __init__.py since it called
get_full_type_name in import-time logic.
