+++ tests/test_storageClass.py
@@ -30,8 +30,7 @@ import os
 import pickle
 import unittest
 
-from lsst.daf.butler import StorageClass, StorageClassConfig, StorageClassFactory
-from lsst.daf.butler.datastore.storage_class_delegate import StorageClassDelegate
+from lsst.daf.butler import StorageClass, StorageClassConfig, StorageClassDelegate, StorageClassFactory
 from lsst.daf.butler.tests import MetricsExample
 from lsst.utils.introspection import get_full_type_name
 
@@ -102,7 +101,7 @@ class StorageClassFactoryTestCase(unittest.TestCase):
         self.assertIn("comp1", scc.components)
         r = repr(scc)
         self.assertIn("comp1", r)
-        self.assertIn("lsst.daf.butler.datastore.storage_class_delegate.StorageClassDelegate", r)
+        self.assertIn("lsst.daf.butler.StorageClassDelegate", r)
 
         # Ensure that we have a delegate
         self.assertIsInstance(scc.delegate(), StorageClassDelegate)
